support for CoURN: automatic creation and identification of root feature; prevent deletion of root feature and outgoing links for root feature; synchronization of names of jucm file, concern, and root feature; visualization of feature model configuration issues in problem view
git-svn-id: svn://cserg0.site.uottawa.ca/projetseg/trunk@2922 1d5a7336-c65e-6947-a78c-d4236001c669

+++ src/seg/jUCMNav/model/commands/delete/DeleteGRLNodeCommand.java
@@ -3,6 +3,7 @@
  */
 package seg.jUCMNav.model.commands.delete;
 
+import fm.Feature;
 import grl.GRLNode;
 import grl.IntentionalElementRef;
 import grl.kpimodel.KPIInformationElementRef;
@@ -12,6 +13,7 @@ import org.eclipse.gef.commands.CompoundCommand;
 import seg.jUCMNav.Messages;
 import seg.jUCMNav.model.commands.delete.internal.PreDeleteUrnModelElementCommand;
 import seg.jUCMNav.model.commands.delete.internal.RemoveURNmodelElementCommand;
+import seg.jUCMNav.model.util.MetadataHelper;
 import seg.jUCMNav.strategies.util.ReusedElementUtil;
 import seg.jUCMNav.views.preferences.DeletePreferences;
 
@@ -33,6 +35,8 @@ public class DeleteGRLNodeCommand extends CompoundCommand {
         setLabel(Messages.getString("DeleteGRLNodeCommand.deleteGrlNode")); //$NON-NLS-1$
         add(new PreDeleteUrnModelElementCommand(ref));
         add(new RemoveURNmodelElementCommand(ref));
+        // wonder if the code of disabling the deletion of the last reference to the
+        // last root feature should display here
     }
 
     /*
@@ -46,13 +50,18 @@ public class DeleteGRLNodeCommand extends CompoundCommand {
             IntentionalElementRef intElement = (IntentionalElementRef) ref;
 
             // Verify if this reference is the only one and it is not a reused element since the
-            // definition should not be deleted (it is found in another model)
-            if (!ReusedElementUtil.isReusedElement(intElement.getDiagram().getUrndefinition().getUrnspec()
-            		.getGrlspec(), intElement.getDef()) && intElement.getDef().getRefs().size() == 1 
-            		&& DeletePreferences.getDeleteDefinition(ref)) {
-                // Add a command to delete the intentional element before execution
-                add(new DeleteIntentionalElementCommand(intElement.getDef()));
-            }
+            String value= MetadataHelper.getMetaData( (intElement).getDef(), "CoURN");  //$NON-NLS-1$
+            
+            boolean rootFeature= (value != null && intElement.getDef() instanceof Feature &&
+         		   value.equalsIgnoreCase("root feature"));  //$NON-NLS-1$
+           
+             if (!ReusedElementUtil.isReusedElement(intElement.getDiagram().getUrndefinition().getUrnspec()
+             		.getGrlspec(), intElement.getDef()) && intElement.getDef().getRefs().size() == 1 
+             		&& DeletePreferences.getDeleteDefinition(ref) && !rootFeature) {
+                 // Add a command to delete the intentional element before execution
+                
+             	add(new DeleteIntentionalElementCommand(intElement.getDef()));
+             }
         } else if (ref instanceof KPIInformationElementRef) {
             KPIInformationElementRef element = (KPIInformationElementRef) ref;
             // Verify if this reference is the only one
@@ -64,4 +73,15 @@ public class DeleteGRLNodeCommand extends CompoundCommand {
 
     }
 
+    public boolean canExecute(){
+    	if (ref instanceof IntentionalElementRef){
+    		 IntentionalElementRef intElement = (IntentionalElementRef) ref;
+    		 String value= MetadataHelper.getMetaData( (intElement).getDef(), "CoURN");
+    		 boolean rootFeature= (value != null && intElement.getDef() instanceof Feature &&
+    	     		   value.equalsIgnoreCase("root feature"));
+    		 if (rootFeature==true && intElement.getDef().getRefs().size()==1)
+    			 return false;
+    	}
+    	return true;
+     }
 }
